// frontend/src/frontend-ts-axios-package/model/search-response-body.ts
/* tslint:disable */
/* eslint-disable */
/**
 * tech blog hub site의 API
 * 카테고리, 회사, 검색어를 기반으로 한 검색 API (페이징 지원)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { ArticleInfo } from './article-info';
// May contain unused imports in some cases
// @ts-ignore
import { PageInfo } from './page-info';

/**
 * 
 * @export
 * @interface SearchResponseBody
 */
export interface SearchResponseBody {
    /**
     * 
     * @type {Array<ArticleInfo>}
     * @memberof SearchResponseBody
     */
    'articleInfos': Array<ArticleInfo>;
    /**
     * 
     * @type {PageInfo}
     * @memberof SearchResponseBody
     */
    'page': PageInfo;
}