// frontend/src/frontend-ts-axios-package/model/article-info.ts
/* tslint:disable */
/* eslint-disable */
/**
 * tech blog hub site의 API
 * 카테고리, 회사, 검색어를 기반으로 한 검색 API (페이징 지원)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ArticleInfo
 */
export interface ArticleInfo {
    /**
     * 
     * @type {string}
     * @memberof ArticleInfo
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof ArticleInfo
     */
    'pubDate': string;
    /**
     * 
     * @type {string}
     * @memberof ArticleInfo
     */
    'companyName': string;
    /**
     * 
     * @type {string}
     * @memberof ArticleInfo
     */
    'url': string;
    /**
     * 
     * @type {string}
     * @memberof ArticleInfo
     */
    'summarizedText': string;
    /**
     * 
     * @type {Array<string>}
     * @memberof ArticleInfo
     */
    'hashtags': Array<string>;
}